+++ packages/di/src/injection/injector.ts
@@ -26,6 +26,13 @@ class _NullInjector implements Injector {
 
 let CURRENT_INJECTOR: Injector = null;
 
+export interface InjectorGetOptions {
+  /**
+   * Do not fall back to parent injector (if one exists for this injector).
+   */
+  self?: boolean;
+}
+
 /**
  * @whatItDoes Injector interface
  * @howToUse
@@ -51,13 +58,19 @@ export abstract class Injector {
   static NULL: Injector = new _NullInjector();
 
   /**
+   * Retrieve the parent injector, if one exists.
+   * Not all injectors support this.
+   */
+  readonly parent?: Injector;
+
+  /**
    * Retrieves an instance from the injector based on the provided token.
    * If not found:
    * - Throws {@link NoProviderError} if no `notFoundValue` that is not equal to
    * Injector.THROW_IF_NOT_FOUND is given
    * - Returns the `notFoundValue` otherwise
    */
-  abstract get<T>(token: Type<T> | InjectionToken<T>, notFoundValue?: T): T;
+  abstract get<T>(token: Type<T> | InjectionToken<T>, notFoundValue?: T, options?: InjectorGetOptions): T;
   /**
    * @deprecated from v4.0.0 use Type<T> or InjectionToken<T>
    * @suppress {duplicate}
@@ -78,9 +91,47 @@ export abstract class Injector {
   }
 }
 
-export function inject<T = unknown>(token: Type<T> | InjectionToken<T>) {
+/**
+ * Type of the options argument to `inject`.
+ *
+ * @publicApi
+ */
+export declare interface InjectOptions {
+  /**
+   * Use optional injection, and return `null` if the requested token is not found.
+   */
+  optional?: boolean;
+  /**
+   * Start injection at the parent of the current injector.
+   */
+  skipSelf?: boolean;
+  /**
+   * Only query the current injector for the token, and don't fall back to the parent injector if
+   * it's not found.
+   */
+  self?: boolean;
+}
+
+/**
+ * Get an injectable value by token from the current injector. This can only be used during creation of an injectable
+ * (ie during class construction). It will throw otherwise.
+ * 
+ * @param token 
+ * @param options 
+ * @returns 
+ */
+export function inject<T = unknown>(token: Type<T> | InjectionToken<T>, options?: InjectOptions) {
   if (CURRENT_INJECTOR === null)
     throw new Error(`inject() can only be called during construction.`);
 
-  return CURRENT_INJECTOR.get(token);
+  let injector = CURRENT_INJECTOR;
+
+  if (options?.skipSelf ?? false)
+    injector = injector.parent ?? injector;
+
+  return CURRENT_INJECTOR.get(
+    token, 
+    options?.optional === true ? undefined : THROW_IF_NOT_FOUND,
+    { self: options?.self ?? false }
+  );
 }